Split up TestWindowFunctions
This makes it easier to find the tests for a given function and will be
necessary as we continue to add functions and tests.
