+++ presto-main/src/test/java/com/facebook/presto/operator/window/AbstractTestWindowFunction.java
@@ -1,0 +1,47 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.facebook.presto.operator.window;
+
+import com.facebook.presto.testing.LocalQueryRunner;
+import com.facebook.presto.testing.MaterializedResult;
+import org.intellij.lang.annotations.Language;
+import org.testng.annotations.AfterClass;
+
+import static com.facebook.presto.SessionTestUtils.TEST_SESSION;
+
+public abstract class AbstractTestWindowFunction
+{
+    protected final LocalQueryRunner queryRunner;
+
+    protected AbstractTestWindowFunction()
+    {
+        queryRunner = new LocalQueryRunner(TEST_SESSION);
+    }
+
+    @AfterClass
+    public void tearDown()
+    {
+        queryRunner.close();
+    }
+
+    protected void assertWindowQuery(@Language("SQL") String sql, MaterializedResult expected)
+    {
+        WindowAssertions.assertWindowQuery(sql, expected, queryRunner);
+    }
+
+    protected void assertWindowQueryWithNulls(@Language("SQL") String sql, MaterializedResult expected)
+    {
+        WindowAssertions.assertWindowQueryWithNulls(sql, expected, queryRunner);
+    }
+}
